requirements.txt
discord.py==2.5.0
Flask==3.0.3
gunicorn==22.0.0
